Split `src/lib.rs` for predicate and contract handling
