+++ tests/json_schema_test_suite_test.rs
@@ -1,6 +1,7 @@
 use std::io::stderr;
 
 use json_schema_test_suite::{json_schema_test_suite, TestCase};
+use json_schema_to_nickel::{predicates::schema_to_predicate, root_schema, wrap_predicate};
 use nickel_lang_core::{eval::cache::lazy::CBNCache, program::Program, term::RichTerm};
 use stringreader::StringReader;
 
@@ -30,13 +31,10 @@ fn translation_typecheck_test(
     test_case: TestCase,
 ) {
     let contract = if test_case.schema.is_object() {
-        json_schema_to_nickel::root_schema(&dbg!(serde_json::from_value(test_case.schema).unwrap()))
+        root_schema(&dbg!(serde_json::from_value(test_case.schema).unwrap()))
     } else {
-        json_schema_to_nickel::wrap_predicate(
-            json_schema_to_nickel::schema_to_predicate(&dbg!(serde_json::from_value(
-                test_case.schema
-            )
-            .unwrap())),
+        wrap_predicate(
+            schema_to_predicate(&dbg!(serde_json::from_value(test_case.schema).unwrap())),
             std::iter::empty(),
         )
     };
